+++ meteor-auth/src/main/java/pers/meteor/auth/service/AuthServiceImpl.java
@@ -90,12 +90,12 @@ public class AuthServiceImpl implements AuthService {
     }
 
     @Override
-    public void sendSmsCode(Integer userId, AuthSmsSendForm smsSendForm) {
+    public void sendSmsCode(Long userId, AuthSmsSendForm smsSendForm) {
 
     }
 
     @Override
-    public void validateSmsCode(Integer userId, AuthSmsValidateForm smsValidateForm) {
+    public void validateSmsCode(Long userId, AuthSmsValidateForm smsValidateForm) {
 
     }
